feat: organizated function
